requirements.txt
groq
google-generativeai
tenacity
python-dotenv
fastapi
uvicorn